Fixed performance regression with multiple GPUs
When using one 1080 Ti, a small (190k seq) dataset could be distilled
in 3 hours per epoch. With 2 GPUs, this was 21h per epoch. By using the
pytorch-lightning DataModule, this is now down to 8h per epoch.

Still not great, but better. Further improvements are likely with the
collate_fn and the organisation of batches.